requirements.txt
torch>=2.6.0
torchvision>=0.21.0
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric>=2.0.0
numpy>=1.26.4
matplotlib>=3.7.0
tqdm>=4.65.0
pillow>=11.1.0
scikit-learn>=1.3.0
nd2reader>=3.3.0
opencv-python>=4.8.0 
cellpose>=4.0.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
opencv-python-headless>=4.5.0
tifffile>=2021.6.6
scipy>=1.7.0
matplotlib>=3.3.0
scikit-image>=0.18.0
tqdm>=4.60.0
tensorboard>=2.5.0
torch-geometric>=2.0.0
pathlib
